# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  model.CourseRecord:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      id:
        type: integer
      material_id:
        type: integer
      progress_percent:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      view_duration:
        type: integer
    type: object
...
  model.Menu:
    properties:
      children:
        description: 关联
        items:
          $ref: '#/definitions/model.Menu'
        type: array
      component:
        type: string
      created_at:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      order_num:
        type: integer
      parent_id:
        type: integer
      path:
        type: string
      permission:
        type: string
      type:
        type: string
      updated_at:
        type: string
      visible:
        type: integer
    type: object
  model.Permission:
    properties:
      action:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      resource:
        type: string
      updated_at:
        type: string
    type: object
  model.Question:
    properties:
      answer:
        type: string
      content:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      exam_id:
        type: integer
      explanation:
        type: string
      id:
        type: integer
      options:
        type: string
      order_num:
        type: integer
      question_type:
        type: string
      score:
        type: number
      updated_at:
        type: string
    type: object
  model.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      menus:
        items:
          $ref: '#/definitions/model.Menu'
        type: array
      name:
        type: string
      permissions:
        description: 关联
        items:
          $ref: '#/definitions/model.Permission'
        type: array
      updated_at:
        type: string
    type: object
  model.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      nickname:
        type: string
      openid:
        type: string
      roles:
        description: 关联
        items:
          $ref: '#/definitions/model.Role'
        type: array
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 学习系统 MVP 后端 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Learn Hub API
  version: "1.0"
paths:
  /admin/permissions:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Permission'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: 获取权限列表
      tags:
      - 管理员
    post:
      parameters:
      - description: 权限数据
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Permission'
              type: object
      security:
      - Bearer: []
      summary: 创建权限
      tags:
      - 管理员
  /admin/roles:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Role'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: 获取角色列表
      tags:
      - 管理员
    post:
      parameters:
      - description: 角色数据
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Role'
              type: object
      security:
      - Bearer: []
      summary: 创建角色
      tags:
      - 管理员
  /admin/roles/{id}:
    delete:
      parameters:
      - description: 角色 ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: 删除角色
      tags:
      - 管理员
    put:
      parameters:
      - description: 角色 ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 更新数据
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Role'
              type: object
      security:
      - Bearer: []
      summary: 更新角色
      tags:
      - 管理员
  /admin/users:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      security:
      - Bearer: []
      summary: 获取用户列表
      tags:
      - 管理员
    post:
      parameters:
      - description: 用户数据
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
      security:
      - Bearer: []
      summary: 创建用户
      tags:
      - 管理员
  /admin/users/{id}:
    delete:
      parameters:
      - description: 用户 ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: 删除用户
      tags:
      - 管理员
    get:
      parameters:
      - description: 用户 ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
      security:
      - Bearer: []
      summary: 获取用户详情
      tags:
      - 管理员
    put:
      parameters:
      - description: 用户 ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 更新数据
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
      security:
      - Bearer: []
      summary: 更新用户
      tags:
      - 管理员
...
  /course-records:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      security:
      - Bearer: []
      summary: 获取学习记录
      tags:
      - 学习记录
  /course-records/{id}:
    get:
      parameters:
      - description: 记录 ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.CourseRecord'
              type: object
      security:
      - Bearer: []
      summary: 获取学习记录详情
      tags:
      - 学习记录
    put:
      parameters:
      - description: 记录 ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 更新数据
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.CourseRecord'
              type: object
      security:
      - Bearer: []
      summary: 更新学习记录
      tags:
      - 学习记录
...
  /menus:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Menu'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: 获取用户菜单
      tags:
      - 菜单
  /questions:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 题型
        in: query
        name: type
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      security:
      - Bearer: []
      summary: 获取题目列表
      tags:
      - 题库
    post:
      parameters:
      - description: 题目数据
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Question'
              type: object
      security:
      - Bearer: []
      summary: 创建题目
      tags:
      - 题库
  /questions/{id}:
    delete:
      parameters:
      - description: 题目 ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: 删除题目
      tags:
      - 题库
    get:
      parameters:
      - description: 题目 ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Question'
              type: object
      security:
      - Bearer: []
      summary: 获取题目详情
      tags:
      - 题库
    put:
      parameters:
      - description: 题目 ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 更新数据
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Question'
              type: object
      security:
      - Bearer: []
      summary: 更新题目
      tags:
      - 题库
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"